sender/requirements.txt
redis==3.1.0
requests==2.21.0
PyYAML==3.13
rq==0.13.0